combine utility functions and classes into renamed utils.py
